xhci: Count link TRBs in free TRBs per ring.
We need to know if we've given the link TRB over to the hardware after
John's patch, so we must count the link TRBs.  This makes it a bit
difficult, since the room_on_ring() function must figure out how many link
TRBs a new transfer is going to use.  Should be simple math though.

Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>