xhci: Handle toggle and cycle bits correctly when expanding rings.
When adding new segment(s) into the ring, make sure the cycle bit of every
TRB in the new segments is opposite of the ring's current cycle bit.  If
this isn't done, the hardware will think it owns all the shiny new TRBs
that software hasn't enqueued to yet.

When adding segments to the end of the ring, make sure to set the toggle
bit to the last added segment, and clear the toggle bit on the segment
we're grafting to.

Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>